requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Trading and data analysis
ccxt>=4.1.0  # For exchange API interactions
ta>=0.10.0   # Technical analysis library
matplotlib>=3.7.0  # For plotting
seaborn>=0.12.0   # For statistical visualizations

# Development tools
pytest>=7.4.0     # For testing
black>=23.7.0     # For code formatting
flake8>=6.1.0     # For linting
pylint>=2.17.0    # For code analysis

# Documentation
sphinx>=7.0.0     # For documentation
sphinx-rtd-theme>=1.3.0  # Documentation theme 